requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
matplotlib>=3.7
jupyter>=1.0
kiteconnect>=5.0
python-dotenv>=1.0
